Backend/servers/subscriptions/pyproject.toml
[tool.poetry]
name = "subscriptions"
version = "0.1.0"
description = ""
authors = ["1twalters01 <1twalters01@gmail.com>"]
readme = "README.rst"

[tool.poetry.scripts]
run = "src.main:application"
activate-venv = "scripts.venv:activate"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.32"
psycopg2-binary = "^2.9.9"
fernet = "^1.0.1"
python-dateutil = "^2.9.0.post0"
cryptography = "^43.0.0"
pyjwt = "^2.9.0"
stripe = "^10.9.0"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
